Moved more things out of core.go
